+++ src/main/java/com/example/shawarmahouse/model/User.java
@@ -24,7 +24,7 @@ public class User {
     @Pattern(regexp = "^\\d{10}$", message = "Phone number must be exactly 10 digits")
     private String phone;
 
-    private long tokens;
+    private Double tokens;
 
 
 }
